Deleted old CD files, reorganized components to be more modular
