+++ components/profile/logoutScreen.js
