+++ components/profile/settingsScreen.js
@@ -21,7 +21,7 @@ const SettingsScreen = ({toggleSettings, logoutModalToggle, changePassModalToggl
       <ScrollView>
         <View style={styles.headerContainer}>
           <ImageBackground
-            source={require('../assets/BOC.profile.cloud.bg.webp')}
+            source={require('../../assets/BOC.profile.cloud.bg.webp')}
             style={{width: 400, minHeight:226}}>
             <View style={styles.headerImage}></View>
             <View style={styles.backButton}>
